+++ app/Http/Controllers/CategoryController.php
@@ -2,8 +2,8 @@
 
 namespace App\Http\Controllers;
 
-use App\Category;
-use App\Product;
+use App\Models\Category;
+use App\Models\Product;
 use Illuminate\Http\Request;
 
 class CategoryController extends Controller
@@ -85,7 +85,7 @@ class CategoryController extends Controller
   /**
    * Display the specified resource.
    *
-   * @param  \App\Category  $category
+   * @param  \App\Models\Category  $category
    * @return \Illuminate\Http\Response
    */
   public function show(Category $category)
@@ -96,7 +96,7 @@ class CategoryController extends Controller
   /**
    * Show the form for editing the specified resource.
    *
-   * @param  \App\Category  $category
+   * @param  \App\Models\Category  $category
    * @return \Illuminate\Http\Response
    */
   public function edit(Category $category)
@@ -108,7 +108,7 @@ class CategoryController extends Controller
    * Update the specified resource in storage.
    *
    * @param  \Illuminate\Http\Request  $request
-   * @param  \App\Category  $category
+   * @param  \App\Models\Category  $category
    * @return \Illuminate\Http\Response
    */
   public function update(Request $request, $id)
@@ -134,7 +134,7 @@ class CategoryController extends Controller
   /**
    * Remove the specified resource from storage.
    *
-   * @param  \App\Category  $category
+   * @param  \App\Models\Category  $category
    * @return \Illuminate\Http\Response
    */
   public function destroy($id)