+++ app/Http/Controllers/ProductController.php
@@ -2,11 +2,11 @@
 
 namespace App\Http\Controllers;
 
-use App\Product;
-use App\Stock;
-use App\Category;
-use Session;
+use App\Models\Category;
+use App\Models\Product;
+use App\Models\Stock;
 use Illuminate\Http\Request;
+use Session;
 
 class ProductController extends Controller
 {
@@ -108,7 +108,7 @@ class ProductController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  \App\Product  $product
+     * @param  \App\Models\Product  $product
      * @return \Illuminate\Http\Response
      */
     public function show(Product $product)
@@ -119,7 +119,7 @@ class ProductController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\Product  $product
+     * @param  \App\Models\Product  $product
      * @return \Illuminate\Http\Response
      */
     public function edit(Product $product)
@@ -131,7 +131,7 @@ class ProductController extends Controller
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Product  $product
+     * @param  \App\Models\Product  $product
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, $id)
@@ -161,7 +161,7 @@ class ProductController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Product  $product
+     * @param  \App\Models\Product  $product
      * @return \Illuminate\Http\Response
      */